requirements.txt
google-api-python-client
python-dotenv
google-genai
youtube-transcript-api